Adding rerun command to cla
This command can be used when the cla check was not run due to GitHub or GitOps issues.
