+++ src/ContributorLicenseAgreement.Core/Handlers/PullRequestHandler.cs
@@ -5,12 +5,10 @@
 
 namespace ContributorLicenseAgreement.Core.Handlers
 {
-    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
     using ContributorLicenseAgreement.Core.Handlers.Helpers;
-    using ContributorLicenseAgreement.Core.Handlers.Model;
     using ContributorLicenseAgreement.Core.Primitives.Data;
     using GitOps.Abstractions;
     using GitOps.Apps.Abstractions.AppEventHandler;
@@ -19,9 +17,7 @@ namespace ContributorLicenseAgreement.Core.Handlers
     using GitOps.Clients.Aad;
     using GitOps.Clients.Ospo;
     using Microsoft.Extensions.Logging;
-    using Polly;
     using Check = ContributorLicenseAgreement.Core.Handlers.Model.Check;
-    using PullRequest = GitOps.Abstractions.PullRequest;
 
     internal class PullRequestHandler : IAppEventHandler
     {
@@ -85,146 +81,11 @@ namespace ContributorLicenseAgreement.Core.Handlers
                 return appOutput;
             }
 
-            if (NeedsLicense(primitive, gitOpsPayload.PullRequest))
-            {
-                logger.LogInformation("License needed for {Sender}", gitOpsPayload.PullRequest.User);
-
-                var hasCla = await HasSignedClaAsync(appOutput, gitOpsPayload, primitive.AutoSignMsftEmployee, primitive.Content);
-
-                appOutput.Comment = await commentHelper.GenerateClaCommentAsync(primitive, gitOpsPayload, hasCla, gitOpsPayload.PullRequest.User);
-
-                var check = new Check
-                {
-                    Sha = gitOpsPayload.PullRequest.Sha,
-                    RepoId = long.Parse(gitOpsPayload.PullRequest.RepositoryId),
-                    InstallationId = gitOpsPayload.PlatformContext.InstallationId
-                };
-
-                await checkHelper.CreateCheckAsync(
-                    gitOpsPayload,
-                    hasCla,
-                    check,
-                    primitive.CheckSummary);
-
-                appOutput.States ??= new States
-                {
-                    StateCollection = new Dictionary<string, object>()
-                };
-
-                appOutput.States.StateCollection.Add(
-                    $"{Constants.Check}-{ClaHelper.GenerateKey(gitOpsPayload.PullRequest.User, primitive.Content)}",
-                    await checkHelper.AddCheckToStatesAsync(gitOpsPayload, check, primitive.Content));
-            }
-            else
-            {
-                logger.LogInformation("No license needed for {Sender}", gitOpsPayload.PullRequest.User);
-                await checkHelper.CreateCheckAsync(
-                    gitOpsPayload,
-                    true,
-                    new Check { Sha = gitOpsPayload.PullRequest.Sha, RepoId = long.Parse(gitOpsPayload.PullRequest.RepositoryId), InstallationId = gitOpsPayload.PlatformContext.InstallationId },
-                    primitive.CheckSummary);
-            }
+            await claHelper.RunCheck(gitOpsPayload, primitive, appOutput);
 
             appOutput.Conclusion = Conclusion.Success;
 
             return appOutput;
-        }
-
-        private static bool NeedsLicense(Cla primitive, PullRequest pullRequest)
-        {
-            return !primitive.BypassUsers.Contains(pullRequest.User)
-                   && !primitive.BypassOrgs.Contains(pullRequest.OrganizationName)
-                   && (pullRequest.Files.Sum(f => f.Changes) >= primitive.MinimalChangeRequired.CodeLines
-                       || pullRequest.Files.Count >= primitive.MinimalChangeRequired.Files);
-        }
-
-        private async Task<bool> HasSignedClaAsync(AppOutput appOutput, GitOpsPayload gitOpsPayload, bool autoSignMsftEmployee, string claLink)
-        {
-            var cla = await appState.ReadState<SignedCla>(ClaHelper.GenerateRetrievalKey(gitOpsPayload.PullRequest.User, claLink));
-
-            if (cla == null || (cla.Employee && cla.MsftMail == null))
-            {
-                var newCla = await TryCreateCla(appOutput, gitOpsPayload, autoSignMsftEmployee, claLink);
-                if (newCla == null)
-                {
-                    if (cla != null)
-                    {
-                        cla = await claHelper.ExpireCla(gitOpsPayload.PullRequest.User, claLink);
-                        appOutput.States = claHelper.GenerateStates(gitOpsPayload.PullRequest.User, claLink, cla);
-                    }
-
-                    return false;
-                }
-
-                cla = newCla;
-            }
-
-            if (!cla.Employee)
-            {
-                var timestamp = System.DateTimeOffset.Now.ToUnixTimeMilliseconds();
-                return cla.Expires == null || cla.Expires > timestamp;
-            }
-
-            if (await IsStillEmployed(cla))
-            {
-                return true;
-            }
-
-            logger.LogInformation("Unable to resolve {Sender} with aad. Trying to re-create cla", gitOpsPayload.PullRequest.User);
-
-            cla = await TryCreateCla(appOutput, gitOpsPayload, autoSignMsftEmployee, claLink);
-
-            if (cla == null || !(await IsStillEmployed(cla)))
-            {
-                cla = await claHelper.ExpireCla(gitOpsPayload.PullRequest.User, claLink);
-                appOutput.States = claHelper.GenerateStates(gitOpsPayload.PullRequest.User, claLink, cla);
-                return false;
-            }
-
-            return await IsStillEmployed(cla);
-        }
-
-        private async Task<bool> IsStillEmployed(SignedCla cla)
-        {
-            ResolvedUser user = null;
-            await Policy
-                .Handle<Exception>()
-                .OrResult<ResolvedUser>(r => r == null)
-                .WaitAndRetryAsync(
-                    3,
-                    retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt)))
-                .ExecuteAsync(async () =>
-                {
-                    user = await aadRequestClient.ResolveUserAsync(cla.MsftMail);
-                    return user;
-                });
-
-            return user is { WasResolved: true };
-        }
-
-        private async Task<SignedCla> TryCreateCla(AppOutput appOutput, GitOpsPayload gitOpsPayload, bool autoSignMsftEmployee, string claLink)
-        {
-            var gitHubUser = gitOpsPayload.PullRequest.User;
-            if (!autoSignMsftEmployee)
-            {
-                return null;
-            }
-
-            var gitHubLink = await gitHubLinkClient.GetLink(gitHubUser);
-            if (gitHubLink?.GitHub == null)
-            {
-                return null;
-            }
-
-            var cla = claHelper.CreateCla(true, gitHubUser, appOutput, "Microsoft", claLink, msftMail: gitHubLink.Aad.UserPrincipalName);
-            logger.LogInformation("CLA signed for GitHub-user: {Cla}", cla);
-            loggingHelper.LogClaSigned(
-                cla,
-                gitOpsPayload.PullRequest.User,
-                gitOpsPayload.PlatformContext.OrganizationName,
-                gitOpsPayload.PlatformContext.RepositoryName,
-                gitOpsPayload.PullRequest.Number);
-            return cla;
         }
     }
 }
